+++ internal/cli/commands.go
@@ -63,6 +63,10 @@ type baseCommand struct {
 	// for defined input variables
 	varFiles []string
 
+	// ignoreMissingVars determines whether variable overrides that do not correspond
+	// to variables defined in the pack should be ignored or produce an error
+	ignoreMissingVars bool
+
 	// autoApproved is true when the user supplies the --auto-approve or -y flag
 	autoApproved bool
 
@@ -280,6 +284,14 @@ func (c *baseCommand) flagSet(bit flagSetBit, f func(*flag.Sets)) *flag.Sets {
 				Completion: complete.PredictOr(complete.PredictFiles("*.var"), complete.PredictFiles("*.hcl")),
 			},
 			Shorthand: "f",
+		})
+
+		f.BoolVar(&flag.BoolVar{
+			Name:    "ignore-missing-vars",
+			Target:  &c.ignoreMissingVars,
+			Default: false,
+			Usage: `Determines whether override variables not present in the
+					pack should be ignored or produce an error.`,
 		})
 
 		f.StringMapVar(&flag.StringMapVar{